Moved some stuff to other files
