+++ src/grafv4/count_modes/line.rs
@@ -1,4 +1,3 @@
-use std::fmt::Display;
 use crate::AppState;
 use crate::grafv4::countable::Countable;
