+++ src/grafv4/graf_zahl.rs
@@ -1,10 +1,7 @@
 use std::ffi::OsStr;
 use std::path::PathBuf;
-use crate::{AppState, Cli, CountMode};
-use crate::grafv4::count_modes::language::LanguageCount;
-use crate::grafv4::countable::{Countable};
-use crate::grafv4::count_modes::line_type::LineTypeCount;
-use crate::grafv4::count_modes::line::LineCount;
+use crate::{AppState, Cli};
+use crate::grafv4::countable::Countable;
 use crate::grafv4::io_reader::{read_dir, read_file};
 
 #[derive(Default, Clone)]
@@ -15,38 +12,26 @@ pub(crate) struct TreeNode {
     pub(crate) errored: bool,
 }
 
-pub fn count_entrypoint(og_path: &PathBuf, state: &AppState, cli: Cli) {
-    let path = &og_path.canonicalize()
-        .expect(&*format!("Absolut Path could not be found for Path: {}", og_path.display()));
+pub(crate) fn generic_count<CountMode: Countable>(state: &AppState, cli: &Cli, unsafe_path: &PathBuf, default_summary: bool, default_per_file: bool) {
+    let path = &unsafe_path.canonicalize()
+        .expect(&*format!("Absolut Path could not be found for Path: {}", unsafe_path.display()));
     assert!(path.is_absolute(), "Received Filepath is not absolut! {}", path.display());
     assert!(path.exists(), "No File/Folder exists at this Path: {}", path.display());
 
-
-    match cli.mode {
-        CountMode::Line => do_count::<LineCount>(state, &cli, path, true, true),
-        CountMode::LOC => do_count::<LineTypeCount>(state, &cli, path, true, true),
-        CountMode::Language => do_count::<LanguageCount>(state, &cli, path, true, false),
-        _ => {
-            println!("This count mode is currently not supported, sorry!")
-        }
-    };
-}
-
-fn do_count<CountMode: Countable>(state: &AppState, cli: &Cli, path: &PathBuf, default_summary: bool, default_per_file: bool) {
     // This unwrap is safe, because all "/.." are resolved when we canonicalize the Path
     let name = path.file_name().unwrap()
         .to_str().expect("Unable to convert File/Folder Name into Unicode!").to_string();
 
     let enable_summary = cli.summary.to_bool_or(default_summary);
     let enable_per_file = cli.per_file.to_bool_or(default_per_file);
-    let count = count_path::<CountMode>(path, state);
+    let count: (TreeNode, CountMode) = count_dir::<CountMode>(path, state);
     if enable_summary {
         count.1.display_summary(name);
         println!();
     }
 
     if enable_per_file {
-        LineCount::display_legend();
+        CountMode::display_legend();
         print_node(count.0, 0, cli.debug, cli.hide_errors);
     }
 }
@@ -61,7 +46,7 @@ fn print_node(node: TreeNode, indent_size: usize, debug: bool, hide_errors: bool
     }
 }
 
-fn count_path<CountMode: Countable>(path: &PathBuf, state: &AppState) -> (TreeNode, CountMode) {
+fn count_dir<CountMode: Countable>(path: &PathBuf, state: &AppState) -> (TreeNode, CountMode) {
     assert!(path.is_absolute(), "Received Filepath is not absolut! {}", path.display());
     assert!(path.exists(), "No File/Folder exists at this Path: {}", path.display());
 
@@ -118,7 +103,7 @@ fn count_file<CountMode: Countable>(path: &PathBuf, name: &String, state: &AppSt
 
 fn count_folder<CountMode: Countable>(path: &PathBuf, state: &AppState, name: &String) -> (TreeNode, CountMode) {
     let members: Vec<(TreeNode, CountMode)> = match read_dir(path) {
-        Ok(v) => v.iter().map(|p| count_path(p, state)).collect(),
+        Ok(v) => v.iter().map(|p| count_dir(p, state)).collect(),
         Err(e) => {
             return (TreeNode {
                 string: format!("{name}/ => [ERR] {e}"),
