+++ src/grafv4/mod.rs
@@ -1,4 +1,5 @@
-pub mod graf_zahl;
+pub(crate) mod count_mode_selector;
+mod graf_zahl;
 mod io_reader;
 mod countable;
 mod count_modes;
